+++ src/Troubleshooting/SettingsDebug.php
@@ -283,6 +283,59 @@ class SettingsDebug implements ServiceInterface, UtilsSettingGlobalInterface
 							],
 						],
 					],
+					[
+						'component' => 'tab',
+						'tabLabel' => \__('Support', 'eightshift-forms'),
+						'tabContent' => [
+							[
+								'component' => 'intro',
+								'introTitle' => \__('Diagnostics', 'eightshift-forms'),
+								'introSubtitle' => '<p>' . \__("Check these settings to make sure your forms work correctly. For more details check <a href='/wp-admin/site-health.php' target='_blank' rel='noopener noreferrer'>Site Health</a>.", 'eightshift-forms') . '</p>',
+							],
+							[
+								'component' => 'card-inline',
+								'cardInlineTitle' => '<a href="https://www.php.net/supported-versions.php" target="_blank" rel="noopener noreferrer">PHP version</a>',
+								'cardInlineSubTitle' => 'Version of PHP running on the server.',
+								'cardInlineRightContent' => \phpversion(),
+							],
+							[
+								'component' => 'card-inline',
+								'cardInlineTitle' => '<a href="https://wordpress.org/documentation/article/wordpress-versions/" target="_blank" rel="noopener noreferrer">WordPress version</a>',
+								'cardInlineSubTitle' => 'Version of WordPress running on the server.',
+								'cardInlineRightContent' => \get_bloginfo('version'),
+							],
+							[
+								'component' => 'card-inline',
+								'cardInlineTitle' => '<a href="https://www.php.net/manual/en/ini.core.php#ini.memory-limit" target="_blank" rel="noopener noreferrer">Memory limit</a>',
+								'cardInlineSubTitle' => 'Maximum amount of memory (in bytes) that a script is allowed to allocate.',
+								'cardInlineRightContent' => (int)\ini_get('memory_limit') <= 64 ? '<span class="error-text">' . \ini_get('memory_limit') . '</span>' : '<span class="success-text">' . \ini_get('memory_limit') . '</span>',
+							],
+							[
+								'component' => 'card-inline',
+								'cardInlineTitle' => '<a href="https://www.php.net/manual/en/info.configuration.php#ini.max-execution-time" target="_blank" rel="noopener noreferrer">Max execution time</a>',
+								'cardInlineSubTitle' => 'Maximum time (in seconds) a script is allowed to run.',
+								'cardInlineRightContent' => (int)\ini_get('max_execution_time') <= 30 ? '<span class="error-text">' . \ini_get('max_execution_time') . '</span>' : '<span class="success-text">' . \ini_get('max_execution_time') . '</span>',
+							],
+							[
+								'component' => 'card-inline',
+								'cardInlineTitle' => '<a href="https://www.php.net/manual/en/info.configuration.php#ini.max-input-vars" target="_blank" rel="noopener noreferrer">Max input vars</a>',
+								'cardInlineSubTitle' => 'Maximum number of input variables that can be used in a single function.',
+								'cardInlineRightContent' => (int)\ini_get('max_input_vars') <= 50 ? '<span class="error-text">' . \ini_get('max_input_vars') . '</span>' : '<span class="success-text">' . \ini_get('max_input_vars') . '</span>',
+							],
+							[
+								'component' => 'card-inline',
+								'cardInlineTitle' => '<a href="https://www.php.net/manual/en/features.file-upload.common-pitfalls.php" target="_blank" rel="noopener noreferrer">Max POST size</a>',
+								'cardInlineSubTitle' => 'Maximum size of POST data that is accepted.',
+								'cardInlineRightContent' => (int)\ini_get('post_max_size') <= 64 ? '<span class="error-text">' . \ini_get('post_max_size') . '</span>' : '<span class="success-text">' . \ini_get('post_max_size') . '</span>',
+							],
+							[
+								'component' => 'card-inline',
+								'cardInlineTitle' => '<a href="https://www.php.net/manual/en/features.file-upload.common-pitfalls.php" target="_blank" rel="noopener noreferrer">Max upload filesize</a>',
+								'cardInlineSubTitle' => 'Maximum allowed file size for file uploads.',
+								'cardInlineRightContent' => (int)\ini_get('upload_max_filesize') <= 64 ? '<span class="error-text">' . \ini_get('upload_max_filesize') . '</span>' : '<span class="success-text">' . \ini_get('upload_max_filesize') . '</span>',
+							],
+						],
+					],
 				],
 			],
 		];