+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [3.0.4]
+
+### Added
+- New support page in the settings for easier access to server configuration issues.
+- Options to send empty fields to entires.
+
+### Changed
+- Block Editor select option will auto close on selection.
+
+### Fixed
+- File upload validation will no longer break the API returns invalid type.
+- Validation pattern will no longer break if value is array.
+
 ## [3.0.3]
 
 ### Added
@@ -238,6 +251,7 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 - Initial production release.
 
+[3.0.4]: https://github.com/infinum/eightshift-forms/compare/3.0.3...3.0.4
 [3.0.3]: https://github.com/infinum/eightshift-forms/compare/3.0.2...3.0.3
 [3.0.2]: https://github.com/infinum/eightshift-forms/compare/3.0.0...3.0.2
 [3.0.0]: https://github.com/infinum/eightshift-forms/compare/2.2.0...3.0.0